test utils added; update failed tests
