+++ biome.json
@@ -25,7 +25,7 @@
 			"bin/**",
 			"playground/**",
 			".nvmrc",
-			"__tests__/**"
+			"test/**"
 		]
 	},
 	"formatter": {
@@ -45,6 +45,9 @@
 			"style": {
 				"useNodejsImportProtocol": "error"
 			},
+			"complexity": {
+        "noForEach": "off"
+      },
 			"correctness": {
 				"useImportExtensions": "off"
 			}
