+++ package.json
@@ -20,13 +20,14 @@
 		"docs:build": "pnpm -r --filter './docs/**' build",
 		"docs:llms": "npx repomix docs/mcpn-docs/content --include '**/*.md' --ignore '**/1.index.md,index.md,*credits*.md,*mcpn*' --output docs/mcpn-docs/public/llms_full.txt",
 		"docs:dev": "pnpm -r --filter './docs/**' dev",
-		"test:dist": "pnpm -r test:dist",
+		"test": "vitest",
 		"lint": "biome lint",
 		"biome:write-fix": "biome check --write --unsafe",
 		"publish-npm": "pnpm -r --filter='./packages/mcpn' run build && tsx scripts/release.ts",
 		"clean": "find . -name 'node_modules' -type d -prune -exec rm -rf {} \\; 2>/dev/null || true",
 		"clean:dist": "find . -name 'dist' -type d -prune -exec rm -rf {} \\; 2>/dev/null || true",
-		"clean:all": "pnpm clean && pnpm clean:dist"
+		"clean:lock": "find . -name 'pnpm-lock.yaml' -type f -delete 2>/dev/null || true",
+		"clean:all": "pnpm clean && pnpm clean:dist && pnpm clean:lock"
 	},
 	"keywords": [
 		"mcpn",
@@ -80,8 +81,7 @@
 		"vitest": "catalog:testing",
 		"vitest-package-exports": "catalog:testing",
 		"wrangler": "^3.1.0",
-		"yaml": "catalog:testing",
-		"yargs": "catalog:cli",
+		
 		"zod": "catalog:prod"
 	},
 	"packageManager": "pnpm@10.6.3",
