+++ packages/cli/test/tool-integration.test.ts
