+++ packages/mcpn/package.json
@@ -49,23 +49,26 @@
 		"cli:dist": "node dist/cli-entry.mjs --preset thinking"
 	},
 	"dependencies": {
+		"@anthropic-ai/sdk": "^0.39.0",
 		"@modelcontextprotocol/sdk": "catalog:prod",
 		"js-yaml": "catalog:prod",
 		"mri": "^1.2.0",
+		"openai": "^4.91.1",
 		"pathe": "^2.0.3",
 		"zod": "catalog:prod"
 	},
 	"devDependencies": {
-		"@types/chai": "catalog:types",
+		"@mcpn/test-utils": "workspace:*",
 		"@types/js-yaml": "catalog:types",
-		"@types/mocha": "catalog:types",
 		"@types/node": "catalog:types",
+		"@types/semver": "catalog:types",
+		"c12": "catalog:utils",
 		"chai": "catalog:testing",
-		
 		"ts-node": "^10.9.2",
 		"tsx": "catalog:cli",
 		"typescript": "catalog:cli",
-		"unbuild": "catalog:cli"
+		"unbuild": "catalog:cli",
+		"vitest": "catalog:testing"
 	},
 	"packageManager": "pnpm@10.6.3"
 }
