+++ packages/mcpn/src/@types/client.ts
@@ -1,3 +1,26 @@
 import type { TemplateParams } from "./common";
 
-// Empty for now, specific types related to client usage can be added later if needed.
+// Type definitions for MCP Test Client interactions
+
+export interface Tool {
+	name: string;
+	description?: string;
+	inputSchema?: any; // Define a more specific type if possible
+}
+
+export interface ListToolsResponse {
+	tools: Tool[];
+	// Add other potential properties like nextCursor if needed
+}
+
+export interface ContentItem {
+	text?: string; // Make text optional
+	type?: string;
+	// Allow other potential properties
+	[key: string]: any;
+}
+
+export interface CallToolResponse {
+	content: ContentItem[];
+	// Add other potential properties if needed
+}
