+++ packages/mcpn/test/test-workflows/.workflows/disable-tool.yaml
@@ -1,2 +1,2 @@
 planner_mode:
-	disabled: true
+  disabled: true
