+++ packages/mcpn/test/test-workflows/.workflows/malformed.yaml
@@ -1,3 +1,3 @@
 this is not valid: yaml:
   - missing colon
-	indentation problem
+  indentation problem
