+++ packages/test-utils/src/index.ts
@@ -1,0 +1,1 @@
+export { McpTestClient } from "./McpTestClient.js";
