+++ pnpm-workspace.yaml
@@ -52,18 +52,13 @@ catalogs:
     unimport: ^4.1.2
     zod: ^3.24.2
   testing:
-    chai: ^4.3.7
-    mocha: ^10.2.0
-    sinon: ^20.0.0
     vitest: ^3.0.8
     vitest-package-exports: ^0.1.1
     yaml: ^2.7.0
   types:
-    '@types/chai': ^4.3.5
     '@types/debug': ^4.1.12
     '@types/fs-extra': ^11.0.4
     '@types/js-yaml': ^4.0.9
-    '@types/mocha': ^10.0.1
     '@types/node': latest
     '@types/semver': ^7.7.0
   utils:
